[PATCH 1/3] drivers: dw dma: dont clear the DONE bit during STOP
Clearing the LL DONE bit tells the DMAC that this LLI descriptor can
be executed again to initiate another transfer. This should always be
done prior to START but not during STOP.

STOP needs all the LLI descriptors DONE to be set so that any LLI load
will stop the DMA state machine.

Signed-off-by: Liam Girdwood <liam.r.girdwood@linux.intel.com>